+++ upload/src/addons/SV/WarningImprovements/XF/Entity/WarningDefinition.php
@@ -120,7 +120,7 @@ class WarningDefinition extends XFCP_WarningDefinition
             $userRepo = Helper::repository(\XF\Repository\User::class);
             $sender = $userRepo->getGuestUser(\XF::phrase('WarningStaff')->render());
 
-            $warningUserId = (int)(\XF::app()->options()->sv_warningimprovements_warning_user ?? 0);
+            $warningUserId = \XF::app()->options()->sv_warningimprovements_warning_user ?? 0;
             if ($warningUserId)
             {
                 $warningStaff = Helper::find(UserEntity::class, $warningUserId);
